pipeline/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
requests==2.31.0
smmap==5.0.1
urllib3==2.1.0
